Refactoring movea
