+++ rpgJavaInterpreter-core/src/main/kotlin/com/smeup/rpgparser/interpreter/internal_interpreter.kt
@@ -412,7 +412,7 @@ class InternalInterpreter(val systemInterface: SystemInterface) : InterpreterCor
                     log(MoveStatemenExecutionLog(this.interpretationContext.currentProgramName, statement, value))
                 }
                 is MoveAStmt -> {
-                    val value = movea(statement.target, statement.expression)
+                    val value = movea(statement.target, statement.expression, this)
                     log(MoveAStatemenExecutionLog(this.interpretationContext.currentProgramName, statement, value))
                 }
                 is MoveLStmt -> {
@@ -1223,45 +1223,6 @@ class InternalInterpreter(val systemInterface: SystemInterface) : InterpreterCor
             MULT_ASSIGNMENT -> assignEachElement(target, eval(MultExpr(target, value)))
             DIVIDE_ASSIGNMENT -> assignEachElement(target, eval(DivExpr(target, value)))
             EXP_ASSIGNMENT -> assignEachElement(target, eval(ExpExpr(target, value)))
-        }
-    }
-
-    private fun movea(target: AssignableExpression, value: Expression): Value {
-        return if (target is DataRefExpr) {
-            moveaFullArray(target, value, 1)
-        } else {
-            require(target is ArrayAccessExpr) {
-                "Result must be an Array element"
-            }
-            moveaFullArray(target.array as DataRefExpr, value, (interpret(target.index) as IntValue).value.toInt())
-        }
-    }
-
-    private fun moveaFullArray(target: DataRefExpr, value: Expression, startIndex: Int): Value {
-        require(target.type() is ArrayType) {
-            "Result must be an Array"
-        }
-        var newValue = interpret(value)
-        return if (value is FigurativeConstantRef) {
-            assign(target, newValue)
-        } else {
-            val realSize = target.type().elementSize() * (target.type().numberOfElements() - startIndex + 1)
-            newValue = newValue.takeFirst(realSize)
-            if (value.type().size < realSize) {
-                newValue = newValue.concatenate(
-                    get(target.variable.referred!!).takeLast((realSize - value.type().size))
-                )
-            }
-            val arrayValue = createArrayValue(target.type(), target.type().numberOfElements()) {
-                if (it < (startIndex - 1)) {
-                    get(target.variable.referred!!).asArray().getElement(it + 1)
-                } else {
-                    val index = it - startIndex + 1
-                    val startValue = (index * target.type().elementSize()) + 1
-                    newValue.take(startValue, startValue + target.type().elementSize() - 1)
-                }
-            }
-            assign(target, arrayValue)
         }
     }
